+++ R/frag.simul.compare.R
@@ -1,17 +1,19 @@
 
 .run.simul <- function(iter, observed.graph, layer.attr, initial.layer, ...){
-  res <- lapply(1:iter, function(i){
+  res <- lapply(seq_len(iter), function(i){
     g <- frag.simul.process(initial.layers = initial.layer,
                             from.observed.graph = observed.graph,
                             observed.layer.attr = layer.attr, ...)
     g <- frag.edges.weighting(g, layer.attr)
     # measure the properties of the graph:
-    inter.layer.e <- igraph::E(g)[ igraph::V(g)[igraph::V(g)$layer == 1] %--% igraph::V(g)[igraph::V(g)$layer == 2]]
+    params <- frag.get.parameters(g, "layer")
+    # inter.layer.e <- igraph::E(g)[ igraph::V(g)[igraph::V(g)$layer == 1] %--% igraph::V(g)[igraph::V(g)$layer == 2]]
     c(
-      "edges" = igraph::gsize(g),
+      "edges" = params$edges,
       "weightsum" = sum(igraph::E(g)$weight),
-      "balance" = c(sort(table(igraph::V(g)$layer))[1] / sum(table(igraph::V(g)$layer)), use.names=FALSE),
-      "disturbance" = length(inter.layer.e) / igraph::gsize(g),
+      "balance" = params$balance,
+      "components.balance" = params$components.balance,
+      "disturbance" = params$disturbance,
       frag.layers.admixture(g, "layer"),
       "cohesion" = rbind(frag.layers.cohesion(g, "layer"))
     )
@@ -30,8 +32,8 @@ frag.simul.compare <- function(graph, layer.attr, iter, summarise=TRUE, ...){
     stop("A logical value is required for the 'summary' parameter.")
   }
   # main funtion:
-  resH1 <- .run.simul(iter, graph, layer.attr, 1, ...)
-  resH2 <- .run.simul(iter, graph, layer.attr, 2, ...)
+  resH1 <- .run.simul(iter, graph, layer.attr, initial.layer = 1, ...)
+  resH2 <- .run.simul(iter, graph, layer.attr, initial.layer = 2, ...)
   
   if(! summarise){
     return(list("h1.data" = resH1, "h2.data" = resH2))